Fetch precalculated files from S3 bucket

+++ src/olinda/data/reference_smiles.py
@@ -13,13 +13,7 @@ from torch.utils.data import DataLoader
 from tqdm import tqdm
 import webdataset as wds
 
-
 from olinda.utils import get_workspace_path
-
-
-reference_download_url = (
-    "https://github.com/ersilia-os/groverfeat/raw/main/data/reference_library.csv"
-)
 
 
 class ReferenceSmilesDM(pl.LightningDataModule):
@@ -59,28 +53,12 @@ class ReferenceSmilesDM(pl.LightningDataModule):
             Path(self.workspace / "reference" / "reference_smiles.csv").is_file()
             is False
         ):
-            ### STORE LOCAL COPY OF FILTERED SMILES
-            ref_path = Path(os.path.join(__file__, "..", ".." , ".." , ".." , "data" , "olinda_reference_library.csv")).resolve()
-            df = pd.read_csv(ref_path)
-            df.to_csv(Path(self.workspace / "reference" / "reference_smiles.csv"), header=False, index=False)
-            # download reference files if not already present
-            """
-            resp = requests.get(reference_download_url, stream=True)
-            total = int(resp.headers.get("content-length", 0))
-            with open(
-                self.workspace / "reference" / "reference_smiles.csv", "wb"
-            ) as file, tqdm(
-                desc="Downloading Reference SMILES data files",
-                total=total,
-                unit="iB",
-                unit_scale=True,
-                unit_divisor=1024,
-            ) as bar:
-                for data in resp.iter_content(chunk_size=1024):
-                    size = file.write(data)
-                    bar.update(size)
-            """
-
+            ref_path = os.path.join(os.path.expanduser("~"), "olinda", "precalculated_descriptors", "olinda_reference_library.csv")
+            # check if reference files not already present
+            if os.path.exists(Path(self.workspace / "reference" / "reference_smiles.csv")) == False or os.path.getsize(ref_path) != os.path.getsize(ref_path):
+                df = pd.read_csv(ref_path)
+                df.to_csv(Path(self.workspace / "reference" / "reference_smiles.csv"), header=False, index=False)
+            
         # Check if processed data files already present
         if (
             Path(self.workspace / "reference" / "reference_smiles.cbor").is_file()
